+++ rn/Teacher/ios/Teacher/SpeedGrader/SubmissionViewer.swift
@@ -23,11 +23,16 @@ struct SubmissionViewer: View {
     let assignment: Assignment
     let submission: Submission
     let fileID: String?
+    @Binding var isPagingEnabled: Bool
 
     @Environment(\.appEnvironment) var env
     @Environment(\.viewController) var controller
 
-    @State var hasOverflow = false
+    @State var webView: WeakReference<Core.CoreWebView> = WeakReference()
+    var hasOverflow: Bool {
+        guard let scrollView = webView.value?.scrollView else { return false }
+        return scrollView.contentSize.width > scrollView.frame.width
+    }
 
     var body: some View {
         switch submission.type {
@@ -56,9 +61,11 @@ struct SubmissionViewer: View {
             let file = submission.attachments?.first { fileID == $0.id } ??
                 submission.attachments?.sorted(by: File.idCompare).first
             if let file = file, let url = file.url, let previewURL = file.previewURL {
-                DocViewer(filename: file.filename, previewURL: previewURL, fallbackURL: url)
+                DocViewer(filename: file.filename, previewURL: previewURL, fallbackURL: url) {
+                    isPagingEnabled = !$0
+                }
             } else if let id = file?.id {
-                FileSubmissionViewer(fileID: id)
+                FileViewer(fileID: id)
             }
         case .online_url:
             URLSubmissionViewer(submission: submission)
@@ -92,86 +99,9 @@ struct SubmissionViewer: View {
     }
 
     func handleSize(webView: Core.CoreWebView, height: CGFloat) {
-        hasOverflow = webView.scrollView.contentSize.width > webView.scrollView.frame.width
-    }
-}
-
-struct URLSubmissionViewer: View {
-    let submission: Submission
-
-    @Environment(\.appEnvironment) var env
-    @Environment(\.viewController) var controller
-
-    @State var image: UIImage?
-    @State var loaded: URL?
-    @State var error: Error?
-
-    var body: some View {
-        ScrollView {
-            VStack(spacing: 0) {
-                HStack {
-                    Text("This submission is a URL to an external page. We've included a snapshot of what the page looked like when it was submitted.")
-                        .font(.regular16).foregroundColor(.textDarkest)
-                    Spacer()
-                }
-                Button(action: {
-                    guard let url = submission.url else { return }
-                    env.loginDelegate?.openExternalURL(url)
-                }, label: {
-                    Text(submission.url?.absoluteString ?? "")
-                        .font(.semibold16).foregroundColor(Color(Brand.shared.linkColor))
-                        .padding(.vertical, 16)
-                })
-                if let file = submission.attachments?.first, let url = file.url {
-                    if let error = error, url == loaded {
-                        Text(error.localizedDescription)
-                            .font(.regular16).foregroundColor(.textDanger)
-                    } else if let image = image, url == loaded {
-                        Button(action: {
-                            guard let id = submission.attachments?.first?.id else { return }
-                            env.router.route(
-                                to: "/files/\(id)",
-                                from: controller,
-                                options: .modal(embedInNav: true, addDoneButton: true)
-                            )
-                        }, label: {
-                            Image(uiImage: image)
-                                .resizable()
-                                .aspectRatio(image.size, contentMode: .fill)
-                                .accessibility(label: Text("URL Preview Image"))
-                        })
-                            .padding(.horizontal, -16)
-                    } else {
-                        CircleProgress().padding(32).onAppear {
-                            env.api.makeRequest(url) { (data, _, error) in
-                                image = data.flatMap { UIImage(data: $0) }
-                                self.error = error
-                                self.loaded = url
-                            }
-                        }
-                    }
-                } else {
-                    Text("Preview Unavailable")
-                        .font(.regular16).foregroundColor(.textDark)
-                }
-            }
-                .multilineTextAlignment(.leading)
-                .padding(16)
-        }
-    }
-}
-
-struct FileSubmissionViewer: UIViewControllerRepresentable {
-    let fileID: String
-
-    func makeUIViewController(context: Self.Context) -> UIViewController { UIViewController() }
-
-    func updateUIViewController(_ uiViewController: UIViewController, context: Self.Context) {
-        let prev = uiViewController.children.first as? FileDetailsViewController
-        if prev?.fileID != fileID {
-            prev?.unembed()
-            let next = FileDetailsViewController.create(context: nil, fileID: fileID)
-            uiViewController.embed(next, in: uiViewController.view)
-        }
+        guard self.webView.value != webView else { return }
+        let ref = WeakReference<Core.CoreWebView>()
+        ref.value = webView
+        self.webView = ref
     }
 }
